refactor: 매치 스케쥴 반환 dto 네이밍 변경
