+++ src/main/java/org/example/oddventure/domain/grid/dto/MatchScheduleDto.java
@@ -1,10 +1,10 @@
-package org.example.oddventure.domain.grid.dto.response;
+package org.example.oddventure.domain.grid.dto;
 
 import java.time.LocalDateTime;
 import lombok.Builder;
 
 @Builder
-public record MatchFetchResponse(
+public record MatchScheduleDto(
         Long fetchId,
         String matchName,
         String teamA,
